legalbot/requirements.txt
transformers>=4.35.0
torch>=2.0.0
faiss-cpu>=1.7.4
langchain>=0.1.0
python-telegram-bot>=20.0
sentence-transformers>=2.2.0
huggingface_hub>=0.19.0
numpy>=1.24.0
langdetect>=1.0.9

